fix: all comments from Yandex

+++ controllers/Cards.js
@@ -2,43 +2,55 @@ const Card = require('../models/Card');
 const BadRequestError = require('../errors/BadRequestError');
 const ForbiddenError = require('../errors/ForbiddenError');
 const NotFoundError = require('../errors/NotFoundError');
+const {
+  STATUS_CREATED,
+  DELETE_ITEM,
+} = require('../utils/constants');
 
 module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
-    .then((card) => {
-      if (!card) {
-        throw new BadRequestError('Введены некорректные данные');
+    .then((card) => res.status(STATUS_CREATED).send({ data: card }))
+    .catch((err) => {
+      if (err.name === 'ValidationError') {
+        next(new BadRequestError('Введены некорректные данные'));
+      } else {
+        next(err);
       }
-      res.status(200).send(card);
-    })
-    .catch(next);
+    });
 };
 
 module.exports.getCards = (req, res, next) => {
   Card.find({})
-    .then((cards) => {
-      if (!cards) {
-        throw new NotFoundError('Список карточек не найден');
-      }
-      res.send(cards);
-    })
+    .populate(['owner', 'likes'])
+    .then((cards) => res.send(cards))
     .catch(next);
 };
 
 module.exports.deleteCard = (req, res, next) => {
-  Card.findByIdAndDelete({ _id: req.params.cardId })
-    .orFail(new NotFoundError('Список карточек не найден'))
-    .then((card) => {
-      if (card.owner.toString() !== req.user._id.toString()) {
-        throw new ForbiddenError('У Вас недостаточно прав, чтобы совершить это действие');
+  const userId = req.user._id;
+
+  Card.findById({ _id: req.params.cardId })
+    .then((data) => {
+      if (!data) {
+        throw new NotFoundError('Невозможно найти');
       }
-      Card.findByIdAndRemove(req.params.cardId)
-        .then((data) => {
-          res.status(200).send(data);
+      if (!data.owner.equals(userId)) {
+        throw new ForbiddenError('Невозможно удалить');
+      }
+      Card.findByIdAndDelete({ _id: req.params.cardId })
+        .orFail(() => new NotFoundError('Невозможно найти'))
+        .then(() => {
+          res.send({ message: DELETE_ITEM });
         });
     })
-    .catch(next);
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        next(new BadRequestError('Введены неккоректные данные'));
+      } else {
+        next(err);
+      }
+    });
 };
 
 module.exports.likeCard = (req, res, next) => {
@@ -48,8 +60,17 @@ module.exports.likeCard = (req, res, next) => {
     { new: true },
   )
     .orFail(new NotFoundError('Указанная карточка не найдена'))
-    .then((likesArray) => res.status(200).send(likesArray))
-    .catch(next);
+    .populate(['owner', 'likes'])
+    .then((card) => {
+      res.send({ data: card });
+    })
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        next(new BadRequestError('Введены некорректные данные'));
+      } else {
+        next(err);
+      }
+    });
 };
 
 module.exports.dislikeCard = (req, res, next) => {
@@ -59,6 +80,6 @@ module.exports.dislikeCard = (req, res, next) => {
     { new: true },
   )
     .orFail(new NotFoundError('Указанная карточка не найдена'))
-    .then((likesArray) => res.status(200).send(likesArray))
+    .then((likesArray) => res.status(STATUS_CREATED).send(likesArray))
     .catch(next);
 };
